requirements.txt
Click==7.0
dominate==2.3.4
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.0
pkg-resources==0.0.0
python-dotenv==0.9.1
visitor==0.1.3
Werkzeug==0.14.1
